+++ src/LionWeb.Core/Core/BaseTypes.cs
@@ -457,7 +457,7 @@ public abstract class NodeBase : ReadableNodeBase<INode>, INode
         return false;
     }
 
-    protected virtual IPartitionCommander? GetPartitionCommander()
+    protected internal virtual IPartitionCommander? GetPartitionCommander()
     {
         INode current = this;
         INode? root = null;
@@ -904,359 +904,6 @@ public abstract class NodeBase : ReadableNodeBase<INode>, INode
                     break;
             }
         }
-    }
-
-    public class SingleContainmentEvent<T> where T : INode
-    {
-        private readonly Containment _containment;
-        private readonly INode _newParent;
-        private readonly T? _newValue;
-        private readonly T? _oldValue;
-        private readonly IPartitionCommander? _partitionCommander;
-
-        private INode? _oldParent;
-        private Containment? _oldContainment;
-        private int _oldIndex;
-
-        public SingleContainmentEvent(Containment containment, NodeBase newParent, T? newValue, T? oldValue)
-        {
-            _containment = containment;
-            _newParent = newParent;
-            _oldValue = oldValue;
-            _newValue = newValue;
-
-            _partitionCommander = newParent.GetPartitionCommander();
-        }
-
-        public void CollectOldData()
-        {
-            if (!IsActive())
-                return;
-
-            if (_newValue == null)
-                return;
-
-            _oldParent = _newValue.GetParent();
-            if (_oldParent == null)
-                return;
-
-            _oldContainment = _oldParent.GetContainmentOf(_newValue);
-            if (_oldContainment == null)
-                return;
-
-            _oldIndex = _oldContainment.Multiple
-                ? M2Extensions.AsNodes<INode>(_oldParent.Get(_oldContainment)).ToList().IndexOf(_newValue)
-                : 0;
-        }
-
-        public void RaiseEvent()
-        {
-            if (!IsActive())
-                return;
-
-            switch (_oldValue, _newValue, _oldParent)
-            {
-                case (null, null, _):
-                case ({ }, { }, _) when Equals(_oldValue, _newValue):
-                    // fall-through
-                    break;
-
-                case ({ }, null, _):
-                    _partitionCommander.DeleteChild(_oldValue, _newParent, _containment, 0);
-                    break;
-
-                case (null, { }, null):
-                    _partitionCommander.AddChild(_newParent, _newValue, _containment, 0);
-                    break;
-
-                case ({ }, { }, null):
-                    _partitionCommander.ReplaceChild(_newValue, _oldValue, _newParent, _containment, 0);
-                    break;
-
-                case (null, { }, { })
-                    when _oldParent == _newParent && _oldContainment != _containment:
-                    _partitionCommander.MoveChildFromOtherContainmentInSameParent(_containment, 0, _newValue,
-                        _newParent, _oldContainment, _oldIndex);
-                    break;
-
-                case ({ }, { }, { })
-                    when _oldParent == _newParent && _oldContainment != _containment:
-                    _partitionCommander.DeleteChild(_oldValue, _newParent, _containment, 0);
-                    _partitionCommander.MoveChildFromOtherContainmentInSameParent(_containment, 0, _newValue,
-                        _newParent, _oldContainment, _oldIndex);
-                    break;
-
-                case ({ }, { }, { })
-                    when _oldParent != _newParent:
-                    _partitionCommander.DeleteChild(_oldValue, _newParent, _containment, 0);
-                    _partitionCommander.MoveChildFromOtherContainment(_newParent, _containment, 0, _newValue,
-                        _oldParent, _oldContainment, _oldIndex);
-                    break;
-
-                case (null, { }, { })
-                    when _oldParent != _newParent:
-                    _partitionCommander.MoveChildFromOtherContainment(_newParent, _containment, 0, _newValue,
-                        _oldParent, _oldContainment, _oldIndex);
-                    break;
-
-                default:
-                    throw new ArgumentException("Unknown state");
-            }
-        }
-
-        [MemberNotNullWhen(true, nameof(_partitionCommander))]
-        private bool IsActive() =>
-            _partitionCommander != null && (_partitionCommander.CanRaiseAddChild() ||
-                                            _partitionCommander.CanRaiseDeleteChild() ||
-                                            _partitionCommander.CanRaiseReplaceChild() ||
-                                            _partitionCommander
-                                                .CanRaiseMoveChildFromOtherContainment() ||
-                                            _partitionCommander
-                                                .CanRaiseMoveChildFromOtherContainmentInSameParent() ||
-                                            _partitionCommander
-                                                .CanRaiseMoveChildInSameContainment());
-    }
-
-    public class AddMultipleContainmentsEvent<T> where T : INode
-    {
-        private readonly Containment _containment;
-        private readonly INode _newParent;
-        private readonly Dictionary<T, Old?> _addedValues;
-        private readonly IPartitionCommander? _partitionCommander;
-
-        private int _newIndex;
-
-        public AddMultipleContainmentsEvent(
-            Containment containment,
-            NodeBase newParent,
-            List<T>? addedValues,
-            List<T> existingValues,
-            int? newIndex = null
-        )
-        {
-            _containment = containment;
-            _newParent = newParent;
-            _addedValues = addedValues?.ToDictionary<T, T, Old?>(k => k, k => null) ?? [];
-            _newIndex = newIndex ?? Math.Max(existingValues.Count - 1, 0);
-
-            _partitionCommander = newParent.GetPartitionCommander();
-        }
-
-        public void CollectOldData()
-        {
-            if (!IsActive())
-                return;
-
-            foreach (T addedValue in _addedValues.Keys.ToList())
-            {
-                var oldParent = addedValue.GetParent();
-                if (oldParent == null)
-                    continue;
-
-                var oldContainment = oldParent.GetContainmentOf(addedValue);
-                if (oldContainment == null)
-                    continue;
-
-                var oldIndex = oldContainment.Multiple
-                    ? M2Extensions.AsNodes<INode>(oldParent.Get(oldContainment)).ToList().IndexOf(addedValue)
-                    : 0;
-
-                _addedValues[addedValue] = new Old(oldParent, oldContainment, oldIndex);
-            }
-        }
-
-        public void RaiseEvent()
-        {
-            if (!IsActive())
-                return;
-
-            foreach ((T? added, Old? old) in _addedValues)
-            {
-                switch (added, old)
-                {
-                    case ({ }, null):
-                        _partitionCommander.AddChild(_newParent, added, _containment, _newIndex);
-                        break;
-
-                    case ({ }, { } o) when o.Parent != _newParent:
-                        _partitionCommander.MoveChildFromOtherContainment(
-                            _newParent,
-                            _containment,
-                            _newIndex,
-                            added,
-                            o.Parent,
-                            o.Containment,
-                            o.Index
-                        );
-                        break;
-
-                    
-                    case ({ }, { } o) when o.Parent == _newParent && o.Containment == _containment && o.Index == _newIndex:
-                        // no-op
-                        break;
-                    
-                    case ({ }, { } o) when o.Parent == _newParent && o.Containment == _containment:
-                        _partitionCommander.MoveChildInSameContainment(
-                            _newIndex,
-                            added,
-                            _newParent,
-                            o.Containment,
-                            o.Index
-                        );
-                        break;
-
-                    case ({ }, { } o) when o.Parent == _newParent && o.Containment != _containment:
-                        _partitionCommander.MoveChildFromOtherContainmentInSameParent(
-                            _containment,
-                            _newIndex,
-                            added,
-                            _newParent,
-                            o.Containment,
-                            o.Index
-                        );
-                        break;
-
-                    default:
-                        throw new ArgumentException("Unknown state");
-                }
-
-                _newIndex++;
-            }
-        }
-
-        [MemberNotNullWhen(true, nameof(_partitionCommander))]
-        private bool IsActive() =>
-            _partitionCommander != null && (_partitionCommander.CanRaiseAddChild() ||
-                                            _partitionCommander
-                                                .CanRaiseMoveChildFromOtherContainment() ||
-                                            _partitionCommander
-                                                .CanRaiseMoveChildFromOtherContainmentInSameParent() ||
-                                            _partitionCommander
-                                                .CanRaiseMoveChildInSameContainment());
-
-        private record struct Old(INode Parent, Containment Containment, int Index);
-    }
-
-    public class SetContainmentEvent<T> where T : INode
-    {
-        private readonly Containment _containment;
-        private readonly INode _newParent;
-        private readonly Dictionary<T, Old?> _setValues;
-        private readonly List<T> _existingValues;
-        private readonly IPartitionCommander? _partitionCommander;
-        private readonly List<IListComparer<T>.Change> _changes = [];
-
-        public SetContainmentEvent(
-            Containment containment,
-            NodeBase newParent,
-            List<T>? setValues,
-            List<T> existingValues
-        )
-        {
-            _containment = containment;
-            _newParent = newParent;
-            _setValues = setValues?.ToDictionary<T, T, Old?>(k => k, k => null) ?? [];
-            _existingValues = existingValues;
-
-            _partitionCommander = newParent.GetPartitionCommander();
-
-            if (!IsActive() || setValues == null)
-                return;
-
-            var listComparer = new StepwiseListComparer<T>(existingValues, setValues);
-            _changes = listComparer.Compare();
-        }
-
-        public void CollectOldData()
-        {
-            if (!IsActive())
-                return;
-
-            foreach (T setValue in _setValues.Keys.ToList())
-            {
-                var oldParent = setValue.GetParent();
-                if (oldParent == null)
-                    continue;
-
-                var oldContainment = oldParent.GetContainmentOf(setValue);
-                if (oldContainment == null)
-                    continue;
-
-                var oldIndex = oldContainment.Multiple
-                    ? M2Extensions.AsNodes<INode>(oldParent.Get(oldContainment)).ToList().IndexOf(setValue)
-                    : 0;
-
-                _setValues[setValue] = new Old(oldParent, oldContainment, oldIndex);
-            }
-        }
-
-        public void RaiseEvent()
-        {
-            if (!IsActive())
-                return;
-
-            foreach (var change in _changes)
-            {
-                switch (change)
-                {
-                    case IListComparer<T>.Added added:
-                        switch (added, _setValues[added.Element])
-                        {
-                            case ({ }, null):
-                                _partitionCommander.AddChild(_newParent, added.Element, _containment, added.RightIndex);
-                                break;
-
-                            case ({ }, { } o) when o.Parent != _newParent:
-                                _partitionCommander.MoveChildFromOtherContainment(
-                                    _newParent,
-                                    _containment,
-                                    added.RightIndex,
-                                    added.Element,
-                                    o.Parent,
-                                    o.Containment,
-                                    o.Index
-                                );
-                                break;
-
-                    
-                            case ({ }, { } o) when o.Parent == _newParent && o.Containment != _containment:
-                                _partitionCommander.MoveChildFromOtherContainmentInSameParent(
-                                    _containment,
-                                    added.RightIndex,
-                                    added.Element,
-                                    _newParent,
-                                    o.Containment,
-                                    o.Index
-                                );
-                                break;
-
-                            default:
-                                throw new ArgumentException("Unknown state");
-                        }
-                        break;
-                    
-                    case IListComparer<T>.Moved moved:
-                        _partitionCommander.MoveChildInSameContainment(moved.RightIndex, moved.LeftElement, _newParent, _containment, moved.LeftIndex);
-                        break;
-                    case IListComparer<T>.Deleted deleted:
-                        _partitionCommander.DeleteChild(deleted.Element, _newParent, _containment, deleted.LeftIndex);
-                        break;
-                }
-            }
-        }
-
-        [MemberNotNullWhen(true, nameof(_partitionCommander))]
-        private bool IsActive() =>
-            _partitionCommander != null && (_partitionCommander.CanRaiseAddChild() ||
-                                            _partitionCommander.CanRaiseDeleteChild() ||
-                                            _partitionCommander
-                                                .CanRaiseMoveChildFromOtherContainment() ||
-                                            _partitionCommander
-                                                .CanRaiseMoveChildFromOtherContainmentInSameParent() ||
-                                            _partitionCommander
-                                                .CanRaiseMoveChildInSameContainment());
-
-        private record struct Old(INode Parent, Containment Containment, int Index);
     }
 
     #endregion
